select changes from my original securities branch
